Rename login-related forms to multi word components
