+++ client/src/components/login/LoginIndex.vue
@@ -1,10 +1,10 @@
 <template>
     <div>
         <template v-if="showLogin">
-            <login :show_welcome_with_login="show_welcome_with_login" :welcome_url="welcome_url" />
+            <login-form :show_welcome_with_login="show_welcome_with_login" :welcome_url="welcome_url" />
         </template>
         <template v-else>
-            <register
+            <register-form
                 :registration_warning_message="registration_warning_message"
                 :mailing_join_addr="mailing_join_addr"
                 :server_mail_configured="server_mail_configured"
@@ -13,8 +13,8 @@
     </div>
 </template>
 <script>
-import Login from "components/login/Login.vue";
-import Register from "components/login/Register.vue";
+import LoginForm from "components/login/LoginForm.vue";
+import RegisterForm from "components/login/RegisterForm.vue";
 import Vue from "vue";
 import BootstrapVue from "bootstrap-vue";
 
@@ -22,8 +22,8 @@ Vue.use(BootstrapVue);
 
 export default {
     components: {
-        login: Login,
-        register: Register,
+        loginForm: LoginForm,
+        registerForm: RegisterForm,
     },
     props: {
         show_welcome_with_login: {
