+++ client/src/components/login/NewUserConfirmation.vue
